Title: 세 수의 합과 평균 구하기, Time: 141ms, Memory: 11MB, Status: Passed - CodeTree

+++ 231214/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -9,7 +9,7 @@ public class Main {
         double c = sc.nextDouble();
 
 
-        System.out.print((int)(n+a+c) + "\n" + (n+c+a)/3);
+        System.out.print((int)(n+a+c) + "\n" + (int)(n+c+a)/3);
 
 
     }
